feat: itsmv4 提单流程 (#257)
* fix approve

* fix 提单

* add push-image

* remove unuse code

* 增加只读字段

* fix lint

* fix lint

* add logs

* fix list biz

* add 租户id

+++ cmd/data-service/service/audit.go
@@ -101,7 +101,7 @@ func (s *Service) ListAudits(ctx context.Context, req *pbds.ListAuditsReq) (*pbd
 				ItsmTicketUrl:     value.Strategy.ItsmTicketUrl,
 				ItsmTicketSn:      value.Strategy.ItsmTicketSn,
 				ItsmTicketStatus:  value.Strategy.ItsmTicketStatus,
-				ItsmTicketStateId: int32(value.Strategy.ItsmTicketStateID),
+				ItsmTicketStateId: value.Strategy.ItsmTicketStateID,
 				Memo:              value.Strategy.Memo,
 			},
 			App: &pbapp.AuditApp{
